* Move more code to `Channel.BasicPublish.cs`
